move constants from doc.go files
