+++ cwf/cloud/go/services/cwf/const.go
@@ -12,3 +12,7 @@
  */
 
 package cwf
+
+const (
+	ServiceName = "cwf"
+)
