+++ feg/cloud/go/services/feg/doc.go
@@ -12,7 +12,3 @@
  */
 
 package feg
-
-const (
-	ServiceName = "feg"
-)
