+++ orc8r/cloud/go/services/orchestrator/doc.go
@@ -12,7 +12,3 @@
  */
 
 package orchestrator
-
-const (
-	ServiceName = "orchestrator"
-)
