+++ orc8r/cloud/go/services/tenants/doc.go
@@ -14,7 +14,7 @@
 package tenants
 
 const (
-	ServiceName          = "TENANTS"
+	ServiceName          = "tenants"
 	DBTableName          = "tenants"
 	TenantInfoType       = "tenant_info"
 	ControlProxyInfoType = "control_proxy_info"
